==> research/04_models.ipynb <==
import os

%pwd

os.chdir("../")

import pandas as pd

data = pd.read_csv("artifacts/data_transformation/train.csv")

data.info()

# set categorical variables

# gender_id , and type as categorical variables
categorical_cols = ["gender_id", "type"]
data[categorical_cols] = data[categorical_cols].astype(str)

data.info()

# Step 1: Separate features and target
X = data.drop(columns="plan_abondonment")
y = data["plan_abondonment"]

# TReat non values for gender_id


from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.ensemble import RandomForestClassifier  # or any other model
from sklearn.feature_extraction import DictVectorizer
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.linear_model import LogisticRegression

categorical_cols = ["gender_id", "type"]

# get numerical variables

X_train = data.drop(columns="plan_abondonment")
y_train = data["plan_abondonment"]

numeric_cols = X_train.select_dtypes(include=["int64", "float64"]).columns.tolist()

# Pipeline

# Step 3: ColumnTransformer for imputing
preprocessor = ColumnTransformer(transformers=[
    ("num", SimpleImputer(strategy="median"), numeric_cols),
    ("scaler", StandardScaler(), numeric_cols),
    ("cat", SimpleImputer(strategy="most_frequent"), categorical_cols)
], remainder="passthrough")  # Keep other columns (if any)


# Step 4: Function to rebuild DataFrame from transformed array
def to_dataframe(X_array):
    return pd.DataFrame(X_array, columns=numeric_cols + categorical_cols)

def array_to_str_keyed_dict_list(X):
    """
    Converts a 2D array into a list of dictionaries
    where keys are string representations of column indices.
    """
    return [dict(zip(map(str, range(X.shape[1])), row)) for row in X]


# Full pipeline
pipeline = Pipeline([
    ("preprocessing", preprocessor),
    ("to_dict", FunctionTransformer(array_to_str_keyed_dict_list)),
    ("vectorize", DictVectorizer()),
    ("clf", LogisticRegression(max_iter=1000))
])

pipeline.fit(X_train, y_train)

# Predict
y_pred = pipeline.predict(X_train)

from sklearn.metrics import accuracy_score

test_data = pd.read_csv("artifacts/data_transformation/test.csv")
test_data[categorical_cols] = test_data[categorical_cols].astype(str)
# Predict on test data
test_X = test_data.drop(columns="plan_abondonment")
test_y = test_data["plan_abondonment"]
y_pred_test = pipeline.predict(test_X)
# Evaluate on test data
test_acc = accuracy_score(test_y, y_pred_test)
print("Test Accuracy:", test_acc)

X_train.head()

X_train.columns

X_train["fraud_score"].unique()

# Random forest

from sklearn.ensemble import RandomForestClassifier

# Full pipeline
pipeline = Pipeline([
    ("preprocessing", preprocessor),
    ("to_dict", FunctionTransformer(array_to_str_keyed_dict_list)),
    ("vectorize", DictVectorizer()),
    ("clf", RandomForestClassifier(random_state=42))
])

pipeline.fit(X_train, y_train)

# Predict
y_pred = pipeline.predict(X_train)

from sklearn.metrics import accuracy_score

acc = accuracy_score(y_train, y_pred)
print("Accuracy:", acc)

test_data = pd.read_csv("artifacts/data_transformation/test.csv")
test_data[categorical_cols] = test_data[categorical_cols].astype(str)
# Predict on test data
test_X = test_data.drop(columns="plan_abondonment")
test_y = test_data["plan_abondonment"]
y_pred_test = pipeline.predict(test_X)
# Evaluate on test data
test_acc = accuracy_score(test_y, y_pred_test)
print("Test Accuracy:", test_acc)

# xgboost classifier

from xgboost import XGBClassifier

# Full pipeline
pipeline = Pipeline([
    ("preprocessing", preprocessor),
    ("to_dict", FunctionTransformer(array_to_str_keyed_dict_list)),
    ("vectorize", DictVectorizer()),
     ('classifier', XGBClassifier())
])

pipeline.fit(X_train, y_train)

# Predict
y_pred = pipeline.predict(X_train)

from sklearn.metrics import accuracy_score

acc = accuracy_score(y_train, y_pred)
print("Accuracy:", acc)

test_data = pd.read_csv("artifacts/data_transformation/test.csv")
test_data[categorical_cols] = test_data[categorical_cols].astype(str)
# Predict on test data
test_X = test_data.drop(columns="plan_abondonment")
test_y = test_data["plan_abondonment"]
y_pred_test = pipeline.predict(test_X)
# Evaluate on test data
test_acc = accuracy_score(test_y, y_pred_test)
print("Test Accuracy:", test_acc)

from sklearn.metrics import (
    accuracy_score, precision_score, recall_score, f1_score,
    roc_auc_score, log_loss, confusion_matrix, classification_report,
    average_precision_score, cohen_kappa_score
)


test_data = pd.read_csv("artifacts/data_transformation/test.csv")
test_data[categorical_cols] = test_data[categorical_cols].astype(str)
# Predict on test data
test_X = test_data.drop(columns="plan_abondonment")
test_y = test_data["plan_abondonment"]
y_pred_test = pipeline.predict(test_X)
# Evaluate on test data

from sklearn.metrics import classification_report, roc_auc_score

print(classification_report(test_y, y_pred_test))
y_proba = pipeline.predict_proba(test_X)[:, 1]  # Get probabilities for the positive class
print("ROC AUC:", roc_auc_score(test_y, y_proba))  # use y_proba for probabilistic output


# modular codes

from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class ModelTrainerConfig:
    root_dir: Path
    train_data_path: Path
    test_data_path: Path
    model_name: str
    model_params: dict
    target_column: str

from mlProject.constants import *
from mlProject.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])


    def get_model_trainer_config(self) -> ModelTrainerConfig:
        config = self.config.model_trainer
        model_params = self.params.XGBoost
        schema =  self.schema.TARGET_COLUMN

        create_directories([config.root_dir])

        model_trainer_config = ModelTrainerConfig(
            root_dir=config.root_dir,
            train_data_path = config.train_data_path,
            test_data_path = config.test_data_path,
            model_name = config.model_name,
            model_params = model_params,
            target_column = schema.name
            
        )

        return model_trainer_config


from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.ensemble import RandomForestClassifier  # or any other model
from sklearn.feature_extraction import DictVectorizer
from sklearn.preprocessing import StandardScaler, FunctionTransformer
from sklearn.linear_model import LogisticRegression
import joblib

class ModelTrainer:
    def __init__(self, config: ModelTrainerConfig):
        self.config = config
    def array_to_str_keyed_dict_list(self, X):
        """
        Converts a 2D array into a list of dictionaries
        where keys are string representations of column indices.
        """
        return [dict(zip(map(str, range(X.shape[1])), row)) for row in X]

    
    def train(self):
        train_data = pd.read_csv(self.config.train_data_path)
        test_data = pd.read_csv(self.config.test_data_path)


        train_x = train_data.drop([self.config.target_column], axis=1)
        test_x = test_data.drop([self.config.target_column], axis=1)
        train_y = train_data[[self.config.target_column]]
        test_y = test_data[[self.config.target_column]]
        
        # gender_id , and type as categorical variables
        categorical_cols = ["gender_id", "type"]
        train_x[categorical_cols] = train_x[categorical_cols].astype(str)
        
        numeric_cols = train_x.select_dtypes(include=["int64", "float64"]).columns.tolist()
        
        # Step 3: ColumnTransformer for imputing
        preprocessor = ColumnTransformer(transformers=[
            ("num", SimpleImputer(strategy="median"), numeric_cols),
            ("scaler", StandardScaler(), numeric_cols),
            ("cat", SimpleImputer(strategy="most_frequent"), categorical_cols)
        ], remainder="passthrough")  # Keep other columns (if any)

        pipeline = Pipeline([
            ("preprocessing", preprocessor),
            ("to_dict", FunctionTransformer(self.array_to_str_keyed_dict_list)),
            ("vectorize", DictVectorizer()),
            ('classifier', XGBClassifier(**self.config.model_params, random_state=42))
        ])

        
        pipeline.fit(train_x, train_y)

        joblib.dump(pipeline, os.path.join(self.config.root_dir, self.config.model_name))



try:
    config = ConfigurationManager()
    model_trainer_config = config.get_model_trainer_config()
    model_trainer_config = ModelTrainer(config=model_trainer_config)
    model_trainer_config.train()
except Exception as e:
    raise e
